Vinay, Rnjn | #1099 | Ensure clear blank diagnosis is called only once
There are 2 triggers to clear blank diagnosis. One is scope destroy, and the
other is the presave handler. Both modify global state in rootScope.
Before upgrade of angular 1.4.9, when consultation save takes place, the
.destroy used to be handled before creation of the new controller.
This behaviour has changed, so a new controller creates a blank diagnosis,
after which .destroy handler kicks in and removes it.

To fix this issue, we have introduced a lock on the controller to call
clearing blank diagnosis exactly once.